requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
bitsandbytes==0.49.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cut-cross-entropy==25.1.1
datasets==4.3.0
decord==0.6.0
diffusers==0.36.0
dill==0.4.0
docstring-parser==0.17.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.2
imageio-ffmpeg==0.6.0
importlib-metadata==8.7.1
jinja2==3.1.6
markupsafe==2.1.5
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==12.0.0
pip==25.0.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pyarrow==22.0.0
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.5.0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0+rocm6.4
torchao==0.15.0
torchaudio==2.9.0+rocm6.4
torchvision==0.24.0+rocm6.4
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
trl==0.24.0
typeguard==4.4.4
typer-slim==0.21.0
typing-extensions==4.15.0
tyro==1.0.3
tzdata==2025.3
unsloth==2025.12.9
unsloth-zoo==2025.12.7
urllib3==2.6.2
uv==0.9.18
wheel==0.45.1
xformers==0.0.33.post1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
